move ClientConfigSingleton to separate file
The big GraphQL client file was getting big, and the ClientConfigSingleton is an independent API from the GraphQL client and merits its own file. This move commit makes it easier to track history for another upcoming refactor.
