+++ lib/shared/src/index.ts
@@ -234,8 +234,8 @@ export {
     SourcegraphGraphQLAPIClient,
     graphqlClient,
 } from './sourcegraph-api/graphql'
+export { ClientConfigSingleton, CodyClientConfig } from './sourcegraph-api/clientConfig'
 export {
-    ClientConfigSingleton,
     addCustomUserAgent,
     customUserAgent,
     isNodeResponse,
